Fixed erros in DARC GPU code for compilation with OpenlCL build.
----
An additional 2 commit messages were squashed into this commit:
--
Fix for opencl compilation errors after merging with master
--
Fixed DARC opencl compilation errors after merging with master
--

Old repository SHA1: 5a767f1ea51033a48441d833ad02be83da048b38

+++ source/src/basic/gpu/GPU.cc
@@ -617,8 +617,8 @@ int GPU::_ExecuteKernel(const char *kernel_name, int total_threads, int max_conc
 	kernelRuntime_ = 0.0;
 
 	do {
-		size_t globalWorkSize = { UPPER_MULTIPLE(total_threads, use_threads) };
-		size_t localWorkSize = { use_threads };
+		size_t globalWorkSize = { static_cast<size_t>(UPPER_MULTIPLE(total_threads, use_threads))};
+		size_t localWorkSize = { static_cast<size_t>(use_threads) };
 
 		errNum_ = clEnqueueNDRangeKernel(this_device().commandQueue, kernel, 1, NULL, &globalWorkSize, &localWorkSize, 0, NULL, &kernelEvent );
